requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==3.0.4
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11